feat: Implement EstatusLookupService for dynamic status ID resolution with caching
docs: Add comprehensive documentation for system purpose, tech stack, and features

docs: Create initial documentation structure in the 'docs' directory

docs: Update README with system overview and feature index

docs: Document cross-cutting concerns including API response wrapper and status lookup service

docs: Analyze administrative access and auditing features in the UI prototype

docs: Detail client and supplier catalog features and API contracts

docs: Outline product catalog management and API integration

docs: Describe purchase order management and associated workflows

docs: Explain installation and session context management in the frontend

docs: Summarize dashboard functionality and API integration

docs: Elaborate on inventory management as a ledger system

docs: Review sales and finance features, highlighting integration gaps

docs: Analyze frontend structure and session management

docs: Define system purpose and stakeholders based on high-level design

docs: Summarize validation analysis for backend and frontend coverage

+++ GesvenApi/DTOs/DTOs.cs
@@ -409,7 +409,7 @@ public class AjusteRespuestaDto
 /// </summary>
 public class MovimientoDto
 {
-    public long MovimientoId { get; set; }
+    public int MovimientoId { get; set; }
     public int InstalacionId { get; set; }
     public int ProductoId { get; set; }
     public string ProductoNombre { get; set; } = string.Empty;
